+++ Runtime/IntegerAnimatorParameter.cs
@@ -1,0 +1,11 @@
+namespace dGameBoy101b.AnimatorParameters
+{
+	public class IntegerAnimatorParameter : AnimatorParameter<int>
+	{
+		public override int Value 
+		{
+			get => this.Animator.GetInteger(this.ParameterName);
+			set => this.Animator.SetInteger(this.ParameterName, value);
+		}
+	}
+}