+++ CHANGELOG.md
@@ -10,4 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
  - AnimatorParameter: An abstract base class which links with an Animator and manipulates a named parameter.
+ - IntegerAnimatorParameter: Manipulates integer parameters in Animators.